requirements.txt
streamlit
pandas
opencv-python
depthai
numpy
pydantic
fastapi
uvicorn
psycopg2-binary
requests
pyyaml